+++ helpers/fm0_decoder.h
@@ -9,7 +9,7 @@ extern volatile unsigned int dt;
 extern volatile unsigned int reading_timer;
 extern volatile unsigned int resultante_tempo;
 
-int fm0_decoder(int tari, unsigned int *payload, unsigned int *n, unsigned short pin_rx,
+int fm0_decoder(int tari, unsigned long long *payload, unsigned int *n, unsigned short pin_rx,
                 unsigned char port_rx, unsigned int timeout);
 
 #endif /* FM0_DECODER_H */
